v1.1 removedcompile time warnings

+++ keyPressActions.cpp
@@ -1,6 +1,6 @@
 #include "myHeaderFiles.h"
 
-int cursorPos, startPos, endPos,winRows;
+size_t cursorPos, startPos, endPos,winRows;
 stack<string> backStack,forStack;
 
 void openDirectory(){
